requirements.txt
numpy
pillow==5.4.1
torch
keras
tensorflow
